Recognize some UMD patterns
It appears that some UMD (Universal Module Definition) patterns
in input code can be treated as AMD.  However, it is necessary
to determine module dependencies and the factory function
from a UMD definition.  Added tests and code to treat UMD as
AMD in this manner.  For information on UMD patterns, see
https://github.com/umdjs/umd/.